requirements.txt
numpy
rasterio
dask
sklearn
matplotlib
geopandas
numpy
dask_geopandas
shapely